[engine] Integrate game and game_loop
