+++ src/engine/mod.rs
@@ -1,3 +1,2 @@
 pub(crate) mod engine;
 pub(crate) mod game;
-mod game_loop;
